refactor(extrusions): moved out & aliased extrusions out of CAG
